chore: moved config schema
