+++ docs/modules/Config.ts.md
@@ -17,6 +17,8 @@ Added in v1.0.0
 - [service](#service)
   - [Config](#config)
   - [Config (interface)](#config-interface)
+  - [ConfigSchema](#configschema)
+  - [PartialConfigSchema](#partialconfigschema)
 
 ---
 
@@ -63,6 +65,84 @@ export interface Config {
   readonly parseCompilerOptions: Record<string, unknown>
   readonly examplesCompilerOptions: Record<string, unknown>
 }
+```
+
+Added in v1.0.0
+
+## ConfigSchema
+
+**Signature**
+
+```ts
+export declare const ConfigSchema: Schema.Schema<
+  {
+    readonly $schema: string
+    readonly projectHomepage: string
+    readonly srcDir: string
+    readonly outDir: string
+    readonly theme: string
+    readonly enableSearch: boolean
+    readonly enforceDescriptions: boolean
+    readonly enforceExamples: boolean
+    readonly enforceVersion: boolean
+    readonly exclude: readonly string[]
+    readonly parseCompilerOptions: string | { readonly [x: string]: unknown }
+    readonly examplesCompilerOptions: string | { readonly [x: string]: unknown }
+  },
+  {
+    readonly $schema: string
+    readonly projectHomepage: string
+    readonly srcDir: string
+    readonly outDir: string
+    readonly theme: string
+    readonly enableSearch: boolean
+    readonly enforceDescriptions: boolean
+    readonly enforceExamples: boolean
+    readonly enforceVersion: boolean
+    readonly exclude: readonly string[]
+    readonly parseCompilerOptions: string | { readonly [x: string]: unknown }
+    readonly examplesCompilerOptions: string | { readonly [x: string]: unknown }
+  }
+>
+```
+
+Added in v1.0.0
+
+## PartialConfigSchema
+
+**Signature**
+
+```ts
+export declare const PartialConfigSchema: Schema.Schema<
+  {
+    readonly $schema?: string
+    readonly projectHomepage?: string
+    readonly srcDir?: string
+    readonly outDir?: string
+    readonly theme?: string
+    readonly enableSearch?: boolean
+    readonly enforceDescriptions?: boolean
+    readonly enforceExamples?: boolean
+    readonly enforceVersion?: boolean
+    readonly exclude?: readonly string[]
+    readonly parseCompilerOptions?: string | { readonly [x: string]: unknown }
+    readonly examplesCompilerOptions?: string | { readonly [x: string]: unknown }
+  },
+  {
+    readonly $schema?: string
+    readonly projectHomepage?: string
+    readonly srcDir?: string
+    readonly outDir?: string
+    readonly theme?: string
+    readonly enableSearch?: boolean
+    readonly enforceDescriptions?: boolean
+    readonly enforceExamples?: boolean
+    readonly enforceVersion?: boolean
+    readonly exclude?: readonly string[]
+    readonly parseCompilerOptions?: string | { readonly [x: string]: unknown }
+    readonly examplesCompilerOptions?: string | { readonly [x: string]: unknown }
+  }
+>
 ```
 
 Added in v1.0.0
