+++ docs/modules/Core.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Core.ts
-nav_order: 5
+nav_order: 4
 parent: Modules
 ---
