+++ docs/modules/Domain.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Domain.ts
-nav_order: 6
+nav_order: 5
 parent: Modules
 ---
