+++ docs/modules/FileSystem.ts.md
@@ -1,6 +1,6 @@
 ---
 title: FileSystem.ts
-nav_order: 7
+nav_order: 6
 parent: Modules
 ---
