+++ docs/modules/Logger.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Logger.ts
-nav_order: 8
+nav_order: 7
 parent: Modules
 ---
