+++ docs/modules/Markdown.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Markdown.ts
-nav_order: 9
+nav_order: 8
 parent: Modules
 ---
