+++ docs/modules/Parser.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Parser.ts
-nav_order: 10
+nav_order: 9
 parent: Modules
 ---
