+++ src/Config.ts
@@ -7,13 +7,36 @@ import chalk from "chalk"
 import { Context, Effect, Layer, Option } from "effect"
 import * as NodePath from "node:path"
 import * as tsconfck from "tsconfck"
-import type { ConfigSchema } from "./ConfigSchema"
-import { PartialConfigSchema } from "./ConfigSchema"
 import * as FileSystem from "./FileSystem"
 import * as Process from "./Process"
 
 const PACKAGE_JSON_FILE_NAME = "package.json"
 const CONFIG_FILE_NAME = "docgen.json"
+
+/**
+ * @category service
+ * @since 1.0.0
+ */
+export const ConfigSchema = Schema.struct({
+  "$schema": Schema.string,
+  projectHomepage: Schema.string,
+  srcDir: Schema.string,
+  outDir: Schema.string,
+  theme: Schema.string,
+  enableSearch: Schema.boolean,
+  enforceDescriptions: Schema.boolean,
+  enforceExamples: Schema.boolean,
+  enforceVersion: Schema.boolean,
+  exclude: Schema.array(Schema.string),
+  parseCompilerOptions: Schema.union(Schema.string, Schema.record(Schema.string, Schema.unknown)),
+  examplesCompilerOptions: Schema.union(Schema.string, Schema.record(Schema.string, Schema.unknown))
+})
+
+/**
+ * @category service
+ * @since 1.0.0
+ */
+export const PartialConfigSchema = Schema.partial(ConfigSchema)
 
 /**
  * @category service
