+++ test/FileSystem.test.ts
@@ -54,7 +54,6 @@ describe("FileSystem", () => {
         Exit.succeed([
           "src/CommandExecutor.ts",
           "src/Config.ts",
-          "src/ConfigSchema.ts",
           "src/Core.ts",
           "src/Domain.ts",
           "src/FileSystem.ts",
@@ -78,7 +77,6 @@ describe("FileSystem", () => {
         Exit.succeed([
           "src/CommandExecutor.ts",
           "src/Config.ts",
-          "src/ConfigSchema.ts",
           "src/Core.ts",
           "src/Domain.ts",
           "src/FileSystem.ts",
